Added animation progress handler that calculates new values for animation targets every frame.

+++ srcs/import.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   import.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: htahvana <htahvana@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: saaltone <saaltone@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/14 13:29:44 by htahvana          #+#    #+#             */
-/*   Updated: 2022/11/24 16:07:47 by htahvana         ###   ########.fr       */
+/*   Updated: 2022/12/05 14:14:07 by saaltone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -114,7 +114,7 @@ static void read_sector(t_app *app, t_exportsector *export, int sectorid, int se
 	(void)sector_count;
 	app->sectors[sectorid].corner_count = export->corner_count;
 	export_to_array(app, export, sectorid);
-	app->sectors[sectorid].light = export->light;
+	app->sectors[sectorid].light = (double)export->light;
 	app->sectors[sectorid].ceil_texture = export->ceil_tex;
 	app->sectors[sectorid].ceil_tex_offset = export->ceil_tex_offset;
 	app->sectors[sectorid].floor_texture = export->floor_tex;
